+++ src/vue/pages/Dashboard.vue
@@ -12,24 +12,30 @@
           :scale="scale"
         />
         <div class="dashboard__actions">
-          <button
-            v-if="isAccountCorporate"
-            class="app__button-raised dashboard__action"
-            @click="createIssuanceFormIsShown = true"
-          >
-            <i class="mdi mdi-plus dashboard__plus-icon" />
-            {{ 'dashboard.create-issuance-lbl' | globalize }}
-          </button>
-          <button
-            v-if="currentAsset"
-            class="app__button-raised dashboard__action"
-            @click="transferFormIsShown = true"
-          >
-            <i class="mdi mdi-send mdi-rotate-315 dashboard__send-icon" />
-            {{
-              'dashboard.send-asset-lbl' | globalize({ asset: currentAsset })
-            }}
-          </button>
+          <!-- eslint-disable-next-line max-len -->
+          <template v-if="getModule().canRenderSubmodule(IssuanceDrawerPseudoModule)">
+            <button
+              class="app__button-raised dashboard__action"
+              @click="createIssuanceFormIsShown = true"
+            >
+              <i class="mdi mdi-plus dashboard__plus-icon" />
+              {{ 'dashboard.create-issuance-lbl' | globalize }}
+            </button>
+          </template>
+
+          <!-- eslint-disable-next-line max-len -->
+          <template v-if="getModule().canRenderSubmodule(TransferDrawerPseudoModule)">
+            <button
+              v-if="currentAsset"
+              class="app__button-raised dashboard__action"
+              @click="transferFormIsShown = true"
+            >
+              <i class="mdi mdi-send mdi-rotate-315 dashboard__send-icon" />
+              {{
+                'dashboard.send-asset-lbl' | globalize({ asset: currentAsset })
+              }}
+            </button>
+          </template>
         </div>
       </div>
       <template v-if="currentAsset">
@@ -42,9 +48,14 @@
             :quote-asset="config.DEFAULT_QUOTE_ASSET"
           />
         </div>
-        <div class="dashboard__activity">
-          <movements-history-module
-            v-if="currentAsset"
+        <div
+          class="dashboard__activity"
+          v-if="getModule().canRenderSubmodule(MovementsHistoryModule) &&
+            currentAsset
+          "
+        >
+          <submodule-importer
+            :submodule="getModule().getSubmodule(MovementsHistoryModule)"
             :asset-code="currentAsset"
             :config="{ horizonURL: config.HORIZON_SERVER }"
             :wallet="wallet"
@@ -70,8 +81,6 @@
 </template>
 
 <script>
-import MovementsHistoryModule from '@modules/movements-history'
-
 import AssetSelector from '@/vue/pages/dashboard/Dashboard.AssetSelector.vue'
 import IssuanceForm from '@/vue/forms/IssuanceForm'
 import Transfer from '@/vue/forms/TransferForm'
@@ -82,6 +91,11 @@ import { vuexTypes } from '@/vuex'
 import Loader from '@/vue/common/Loader'
 import config from '@/config'
 import Drawer from '@/vue/common/Drawer'
+import { MovementsHistoryModule } from '@/vue/modules/movements-history/module'
+import SubmoduleImporter from '@/modules-arch/submodule-importer'
+
+import { IssuanceDrawerPseudoModule } from '@/modules-arch/pseudo-modules/issuance-drawer-pseudo-module'
+import { TransferDrawerPseudoModule } from '@/modules-arch/pseudo-modules/transfer-drawer-pseudo-module'
 
 export default {
   name: 'dashboard',
@@ -89,10 +103,10 @@ export default {
     AssetSelector,
     IssuanceForm,
     Transfer,
-    MovementsHistoryModule,
     Chart,
     Loader,
     Drawer,
+    SubmoduleImporter,
   },
   data: () => ({
     currentAsset: null,
@@ -102,6 +116,9 @@ export default {
     showDrawer: false,
     scale: 'month',
     config,
+    MovementsHistoryModule,
+    IssuanceDrawerPseudoModule,
+    TransferDrawerPseudoModule,
   }),
   computed: {
     ...mapGetters([
@@ -187,7 +204,7 @@ export default {
 
 .dashboard__action {
   &:not(:first-child) {
-    margin-left: .8rem;
+    margin-left: 0.8rem;
   }
 }