+++ src/vue/pages/Settings.vue
@@ -3,12 +3,14 @@
     <top-bar>
       <template slot="main">
         <router-link
+          v-if="getModule().canRenderSubmodule(VerificationPageModule)"
           v-ripple
           :to="vueRoutes.verification"
         >
           <span>{{ verificationTabId | globalize }}</span>
         </router-link>
         <router-link
+          v-if="getModule().canRenderSubmodule(SecurityPageModule)"
           v-ripple
           :to="vueRoutes.security"
         >
@@ -29,6 +31,8 @@ import { vuexTypes } from '@/vuex'
 import { mapGetters } from 'vuex'
 
 import { REQUEST_STATES_STR } from '@/js/const/request-states.const'
+import { VerificationPageModule } from '@/vue/pages/verification-page-module'
+import { SecurityPageModule } from '@/vue/pages/security-page-module'
 
 export default {
   name: 'settings',
@@ -37,6 +41,8 @@ export default {
   },
   data: _ => ({
     vueRoutes,
+    VerificationPageModule,
+    SecurityPageModule,
   }),
   computed: {
     ...mapGetters({